# GreenGearAPI/public/swagger.json
{
	"openapi": "3.0.3",
	"info": {
		"title": "GreenGear API",
		"version": "1.0.0",
		"contact": {}
	},
	"servers": [
		{
			"url": "http://localhost"
		}
	],
	"paths": {
...
		"/api/gardens": {
			"get": {
				"tags": [
					"Gardens"
				],
				"summary": "Get All Gardens",
				"description": "Get All Gardens",
				"operationId": "getAllGardens",
				"responses": {
					"200": {
						"description": ""
					}
				}
			},
			"post": {
				"tags": [
					"Gardens"
				],
				"summary": "Create Garden",
				"description": "Create Garden",
				"operationId": "createGarden",
				"requestBody": {
					"content": {
						"application/json": {
							"schema": {
								"type": "object",
								"properties": {
									"garden_name": {
										"type": "string",
										"example": "My Garden 2"
									}
								}
							},
							"example": {
								"garden_name": "My Garden 2"
							}
						}
					}
				},
				"responses": {
					"200": {
						"description": ""
					}
				}
			}
		},
		"/api/gardens/{garden_id}": {
			"get": {
				"tags": [
					"Gardens"
				],
				"summary": "Get Garden By ID",
				"description": "Get Garden By ID",
				"operationId": "getGardenById",
				"responses": {
					"200": {
						"description": ""
					}
				}
			},
			"put": {
				"tags": [
					"Gardens"
				],
				"summary": "Update Garden",
				"description": "Update Garden",
				"operationId": "updateGarden",
				"requestBody": {
					"content": {
						"application/json": {
							"schema": {
								"type": "object",
								"properties": {
									"garden_name": {
										"type": "string",
										"example": "My Garden2"
									},
									"userId": {
										"type": "number",
										"example": 2
									}
								}
							},
							"example": {
								"garden_name": "My Garden2",
								"userId": 2
							}
						}
					}
				},
				"responses": {
					"200": {
						"description": ""
					}
				}
			},
			"delete": {
				"tags": [
					"Gardens"
				],
				"summary": "Delete Garden",
				"description": "Delete Garden",
				"operationId": "deleteGarden",
				"responses": {
					"200": {
						"description": ""
					}
				}
			},
			"parameters": [
				{
					"name": "garden_id",
					"in": "path",
					"required": true,
					"schema": {
						"type": "string",
						"example": ""
					}
				}
			]
		},
...
		"/api/plants/garden/{gardenId}": {
			"get": {
				"tags": [
					"Plants"
				],
				"summary": "Get all Plants from Garden",
				"description": "Get all Plants from Garden",
				"operationId": "getAllPlantsFromGarden",
				"responses": {
					"200": {
						"description": ""
					}
				}
			},
			"parameters": [
				{
					"name": "gardenId",
					"in": "path",
					"required": true,
					"schema": {
						"type": "string",
						"example": "1"
					}
				}
			]
		},
...
		"/api/hub/": {
			"post": {
				"tags": [
					"Hub"
				],
				"summary": "Create Hub",
				"description": "Create Hub",
				"operationId": "createHub",
				"requestBody": {
					"content": {
						"application/json": {
							"schema": {
								"type": "object",
								"properties": {
									"garden_id": {
										"type": "number",
										"example": 1
									},
									"hub_name": {
										"type": "string",
										"example": "Hub#2"
									},
									"hub_token": {
										"type": "string",
										"example": "test_token"
									},
									"hub_version": {
										"type": "string",
										"example": "0.0.1"
									}
								}
							},
							"example": {
								"garden_id": 1,
								"hub_name": "Hub#2",
								"hub_token": "test_token",
								"hub_version": "0.0.1"
							}
						}
					}
				},
				"responses": {
					"200": {
						"description": ""
					}
				}
			}
		},
...
		"/api/hub/{hub_id}": {
			"get": {
				"tags": [
					"Hub"
				],
				"summary": "Get Hub by ID",
				"description": "Get Hub by ID",
				"operationId": "getHubById",
				"requestBody": {
					"content": {
						"text/plain": {
							"example": ""
						}
					}
				},
				"responses": {
					"200": {
						"description": ""
					}
				}
			},
			"put": {
				"tags": [
					"Hub"
				],
				"summary": "Update Hub",
				"description": "Update Hub",
				"operationId": "updateHub",
				"requestBody": {
					"content": {
						"application/json": {
							"schema": {
								"type": "object",
								"properties": {
									"hub_name": {
										"type": "string",
										"example": "Hub#2qwe"
									},
									"hub_token": {
										"type": "string",
										"example": "test_token"
									},
									"hub_version": {
										"type": "string",
										"example": "0.0.1"
									}
								}
							},
							"example": {
								"hub_name": "Hub#2qwe",
								"hub_token": "test_token",
								"hub_version": "0.0.1"
							}
						}
					}
				},
				"responses": {
					"200": {
						"description": ""
					}
				}
			},
			"delete": {
				"tags": [
					"Hub"
				],
				"summary": "Delete Hub",
				"description": "Delete Hub",
				"operationId": "deleteHub",
				"requestBody": {
					"content": {
						"text/plain": {
							"example": ""
						}
					}
				},
				"responses": {
					"200": {
						"description": ""
					}
				}
			},
			"parameters": [
				{
					"name": "hub_id",
					"in": "path",
					"required": true,
					"schema": {
						"type": "string",
						"example": ""
					}
				}
			]
		},
		"/api/hub/{hub_id}/sensors": {
			"get": {
				"tags": [
					"Hub"
				],
				"summary": "Get Hub Sensors",
				"description": "Get Hub Sensors",
				"operationId": "getHubSensors",
				"requestBody": {
					"content": {
						"text/plain": {
							"example": ""
						}
					}
				},
				"responses": {
					"200": {
						"description": ""
					}
				}
			},
			"parameters": [
				{
					"name": "hub_id",
					"in": "path",
					"required": true,
					"schema": {
						"type": "string",
						"example": "1"
					}
				}
			]
		},
		"/api/hub/{hub_id}/gardens": {
			"get": {
				"tags": [
					"Hub"
				],
				"summary": "Get Hub Gardens",
				"description": "Get Hub Gardens",
				"operationId": "getHubGardens",
				"requestBody": {
					"content": {
						"text/plain": {
							"example": ""
						}
					}
				},
				"responses": {
					"200": {
						"description": ""
					}
				}
			},
			"parameters": [
				{
					"name": "hub_id",
					"in": "path",
					"required": true,
					"schema": {
						"type": "string",
						"example": "1"
					}
				}
			]
		},
		"/api/hub/link-garden": {
			"put": {
				"tags": [
					"Hub"
				],
				"summary": "Link Hub to Garden",
				"description": "Link Hub to Garden",
				"operationId": "linkHubToGarden",
				"requestBody": {
					"content": {
						"application/json": {
							"schema": {
								"type": "object",
								"properties": {
									"garden_id": {
										"type": "string",
										"example": "2"
									},
									"hub_id": {
										"type": "string",
										"example": "1"
									}
								}
							},
							"example": {
								"garden_id": "2",
								"hub_id": "1"
							}
						}
					}
				},
				"responses": {
					"200": {
						"description": ""
					}
				}
			}
		}
	},
...
}